Separate platform-specifics from secure-e2ee characteristics
